Correções e ajustes

+++ comandos/manutencao/menu_adm.js
@@ -6,7 +6,10 @@ module.exports = {
     permissions: [ "ADMINISTRATOR" ],
     execute(client, message, args) {
 
-        message.delete();
+        const permissions = message.channel.permissionsFor(message.client.user);
+
+        if(permissions.has("MANAGE_MESSAGES")) // Permissão para gerenciar mensagens
+            message.delete();
 
         const { MessageEmbed } = require('discord.js');
